requirements.txt
paho-mqtt
rshell
micropython-esp32-stubs>=1.20.0.*
